+++ nakadion/src/components/committer/mod.rs
@@ -347,10 +347,12 @@ impl StdError for CommitError {
 impl fmt::Display for CommitError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match (self.context.as_ref(), self.source.as_ref()) {
-            (None, None) =>  write!(f, "{}", self.kind),
-            (None, Some(source)) =>  write!(f, "{} - Caused by: {}", self.kind, source),
-            (Some(context), None) =>  write!(f, "{} - {}", self.kind, context),
-            (Some(context), Some(source)) =>  write!(f, "{} - {} - Caused by: {}", self.kind, context, source),
+            (None, None) => write!(f, "{}", self.kind),
+            (None, Some(source)) => write!(f, "{} - Caused by: {}", self.kind, source),
+            (Some(context), None) => write!(f, "{} - {}", self.kind, context),
+            (Some(context), Some(source)) => {
+                write!(f, "{} - {} - Caused by: {}", self.kind, context, source)
+            }
         }
     }
 }